use random centroids
this avoids a bug. Imagine an almost-entirely red picture. Then all the points would get
classified into the red centroid, and we don't know what to do with the other two centroids because
there are no points in them.

If that happens with our random centroids, just try new centroids. Once we get any pixels in all
three centroids, we're golden.